@@ -1,16 +1,16 @@
 name: hydrated_triple
 description: A mixin which automatically persists and restores store state. 
-version: 0.0.3
+version: 1.0.0
 homepage: https://github.com/Flutterando/triple_pattern
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   shared_preferences: ^2.0.6
-  triple: ">=1.4.0 <2.0.0"
+  triple: ">=2.0.0 <3.0.0"
 
 dev_dependencies:
   flutter_test:
