@@ -1,6 +1,6 @@
 name: flutter_triple
 description: Implementation of the Segmented State Pattern (SSP) nicknamed Triple.
-version: 1.3.0
+version: 2.0.0
 
 repository: https://github.com/Flutterando/triple_pattern/tree/master/stores/flutter_triple
 issue_tracker: https://github.com/Flutterando/triple_pattern/issues
@@ -8,13 +8,14 @@ homepage: https://triple.flutterando.com.br/
 documentation: https://triple.flutterando.com.br/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  triple: ">=1.5.0+1 <2.0.0"
+  triple: ">=2.0.0 <3.0.0"
   collection: ^1.15.0
   flutter:
     sdk: flutter
+  rx_notifier: ">=2.2.0+2 <3.0.0"
 
 dev_dependencies:
   flutter_test:
