@@ -1,23 +1,21 @@
 name: mobx_triple
 description: Segmented State Pattern for MobX
-version: 1.1.3+1
+version: 2.0.0
 homepage: https://github.com/Flutterando/triple_pattern
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   meta: ">=1.3.0 <2.0.0"
   mobx: ">=2.0.5 <3.0.0"
-  triple: ">=1.3.3 <2.0.0"
+  flutter_triple: ">=2.0.0 <3.0.0"
+  triple: ^2.0.0
+  rx_notifier: ^2.2.0+2
 
 dev_dependencies:
   test: ^1.16.5
-  flutterando_analysis: ^0.0.1
+  flutterando_analysis: ^0.0.2
 
 
-
-# dependency_overrides:
-#   triple: 
-#     path: ../triple
 
