@@ -1,6 +1,6 @@
 name: triple_test
 description: Test Helper for Store of flutter_triple with Mocktail.
-version: 1.0.0+1
+version: 2.0.0
 
 repository: https://github.com/Flutterando/triple_pattern
 issue_tracker: https://github.com/Flutterando/triple_pattern/issues
@@ -8,16 +8,16 @@ homepage: https://triple.flutterando.com.br/
 documentation: https://triple.flutterando.com.br/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  triple: ">=1.0.0 <2.0.0"
-  mocktail: ">=0.1.1 <1.0.0"
+  triple: ">=2.0.0 <3.0.0"
+  mocktail: ">=0.3.0 <1.0.0"
   meta: ">=1.3.0 <2.0.0"
   test: ">=1.16.0 <2.0.0"
 
 dev_dependencies:
-  flutterando_analysis: ^0.0.1
+  flutterando_analysis: ^0.0.2
 
 #   flutter_test:
 #     sdk: flutter
