@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_triple/flutter_triple.dart';
 
 import 'counter.dart';
-import 'package:flutter_triple/flutter_triple.dart';
 
 void main() {
   runApp(MyApp());
@@ -31,6 +31,20 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   final counter = CounterStore();
+  late Disposer disposer;
+
+  @override
+  void initState() {
+    super.initState();
+
+    disposer = counter.observer(onState: print);
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    disposer();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -49,7 +63,7 @@ class _MyHomePageState extends State<MyHomePage> {
         ],
       ),
       body: Center(
-        child: ScopedConsumer<CounterStore, Exception, int>(
+        child: ScopedConsumer<CounterStore, int>(
           store: counter,
           onLoadingListener: (context, isLoading) {},
           onErrorListener: (context, error) {
@@ -74,14 +88,13 @@ class _MyHomePageState extends State<MyHomePage> {
           },
         ),
       ),
-      floatingActionButton: TripleBuilder<CounterStore, Exception, int>(
+      floatingActionButton: TripleBuilder<CounterStore, int>(
         store: counter,
         builder: (context, triple) {
           return FloatingActionButton(
             onPressed: triple.isLoading ? null : counter.increment,
             tooltip: triple.isLoading ? 'no-active' : 'Increment',
-            backgroundColor:
-                triple.isLoading ? Colors.grey : Theme.of(context).primaryColor,
+            backgroundColor: triple.isLoading ? Colors.grey : Theme.of(context).primaryColor,
             child: Icon(Icons.add),
           );
         },
