@@ -1,9 +1,9 @@
 library flutter_triple;
 
+export 'package:rx_notifier/rx_notifier.dart';
 export 'package:triple/triple.dart';
 
-export 'src/stores/notifier_store.dart';
-export 'src/stores/stream_store.dart';
+export 'src/stores/store.dart';
 export 'src/widgets/scoped_builder.dart';
 export 'src/widgets/scoped_consumer.dart';
 export 'src/widgets/scoped_listener.dart';
