@@ -6,15 +6,14 @@ import 'package:triple/triple.dart';
 
 ///[TripleConsumer] class it's the type <TStore extends Store<TError, TState>,
 ///TError extends Object, TState extends Object>
-class TripleConsumer<TStore extends Store<TError, TState>,
-    TError extends Object, TState extends Object> extends StatefulWidget {
+class TripleConsumer<TStore extends BaseStore<TState>, TError extends Object, TState extends Object> extends StatefulWidget {
   ///The Function [builder] it's the type [Widget] and receive
   ///the params context it`s the type [BuildContext] and triple it's
   ///the type Triple<TError, TState>
 
   final Widget Function(
     BuildContext context,
-    Triple<TError, TState> triple,
+    Triple<TState> triple,
   ) builder;
 
   ///The Function [listener] it's the type void and receive
@@ -22,17 +21,17 @@ class TripleConsumer<TStore extends Store<TError, TState>,
   ///the type Triple<TError, TState>
   final void Function(
     BuildContext context,
-    Triple<TError, TState> triple,
+    Triple<TState> triple,
   ) listener;
 
   ///The Function [filter] it's the type [bool] and receive the
   ///param triple it`s the type Triple<TError, TState>
 
-  final bool Function(Triple<TError, TState> triple)? filter;
+  final bool Function(Triple<TState> triple)? filter;
 
   ///The Function [distinct] it's the type [dynamic] and receive the
   ///param state it`s the type Triple<TError, TState>
-  final dynamic Function(Triple<TError, TState> state)? distinct;
+  final dynamic Function(Triple<TState> state)? distinct;
 
   ///[store] it's the type [TStore]
   final TStore? store;
@@ -48,13 +47,10 @@ class TripleConsumer<TStore extends Store<TError, TState>,
   }) : super(key: key);
 
   @override
-  _TripleConsumerState<TStore, TError, TState> createState() =>
-      _TripleConsumerState<TStore, TError, TState>();
+  _TripleConsumerState<TStore, TError, TState> createState() => _TripleConsumerState<TStore, TError, TState>();
 }
 
-class _TripleConsumerState<TStore extends Store<TError, TState>,
-        TError extends Object, TState extends Object>
-    extends State<TripleConsumer<TStore, TError, TState>> {
+class _TripleConsumerState<TStore extends BaseStore<TState>, TError extends Object, TState extends Object> extends State<TripleConsumer<TStore, TError, TState>> {
   Widget? child;
 
   var _distinct;
