@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_triple/flutter_triple.dart';
+
 import '../mocks/mocks.dart';
 
 void main() {
@@ -14,7 +15,7 @@ void main() {
     testWidgets('calls builder when state changes', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text('state ${triple.state}'),
           ),
@@ -35,7 +36,7 @@ void main() {
     testWidgets('calls builder when an error is emitted', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text('${triple.error}'),
           ),
@@ -52,7 +53,7 @@ void main() {
     testWidgets('calls builder when loading is emitted', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => const Text('loading'),
           ),
@@ -65,11 +66,10 @@ void main() {
       expect(find.text('loading'), findsOneWidget);
     });
 
-    testWidgets('calls builder when state is emitted with filter',
-        (tester) async {
+    testWidgets('calls builder when state is emitted with filter', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text(triple.state.toString()),
             filter: (triple) => triple.state <= 1,
@@ -85,11 +85,10 @@ void main() {
       expect(find.text('1'), findsOneWidget);
     });
 
-    testWidgets('calls builder when state is emitted with distincted',
-        (tester) async {
+    testWidgets('calls builder when state is emitted with distincted', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text(triple.state.toString()),
             distinct: (triple) => triple.state < 1,
@@ -105,11 +104,10 @@ void main() {
       expect(find.text('1'), findsOneWidget);
     });
 
-    testWidgets('calls builder when state is emitted with distincted',
-        (tester) async {
+    testWidgets('calls builder when state is emitted with distincted', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text(triple.state.toString()),
             // ignore: avoid_redundant_argument_values
@@ -129,7 +127,7 @@ void main() {
     testWidgets('calls builder when load and state changes', (tester) async {
       await tester.pumpWidget(
         MockWidget(
-          child: TripleBuilder<MockStore, String, int>(
+          child: TripleBuilder<MockStore, int>(
             store: store,
             builder: (context, triple) => Text('state ${triple.state}'),
           ),
