@@ -1,6 +1,6 @@
 import 'package:mobx_triple/mobx_triple.dart';
 
-class Counter extends MobXStore<Exception, int> with MementoMixin {
+class Counter extends MobXStore<int> with MementoMixin {
   Counter() : super(0);
 
   increment() {
