@@ -2,7 +2,7 @@
 
 import 'package:meta/meta.dart';
 
-import 'store.dart';
+import 'base_store.dart';
 
 class _MutableFlags {
   bool hasInitiated = false;
@@ -49,13 +49,7 @@ class MemoryHydratedDelegate implements HydratedDelegate {
   @override
   Future save(String key, value) async {
     assert(
-      value is num ||
-          value is String ||
-          value is bool ||
-          value is List<String> ||
-          value is Map ||
-          value is Set ||
-          value is Serializable,
+      value is num || value is String || value is bool || value is List<String> || value is Map || value is Set || value is Serializable,
       'not valid value',
     );
     _cachedValue = value;
@@ -63,7 +57,7 @@ class MemoryHydratedDelegate implements HydratedDelegate {
 }
 
 ///[HydratedMixin] mixin
-mixin HydratedMixin<Error extends Object, State extends Object> on Store<Error, State> {
+mixin HydratedMixin<State> on BaseStore<State> {
   final _flags = _MutableFlags();
 
   ///[keyName] it's a get and it's the type [String]
@@ -80,7 +74,7 @@ mixin HydratedMixin<Error extends Object, State extends Object> on Store<Error, 
     } else {
       _delegate.save(keyName, newState);
     }
-    super.update(newState, force: force);
+    super.update(newState, force: true);
   }
 
   @override
