@@ -3,10 +3,10 @@
 /// More dartdocs go here.
 library triple;
 
+export 'src/base_store.dart';
 export 'src/either_adapter.dart';
 export 'src/hydrated_mixin.dart';
 export 'src/memento_mixin.dart';
 export 'src/models/dispatched_triple.dart';
 export 'src/models/triple_model.dart';
 export 'src/selectors.dart';
-export 'src/store.dart';
